Enable EventTrigger webhook
We built out a webhook for the EventTrigger CRD, but we didn't include
it within the webhook config file. This hooks it up so that
create/update for EventTrigger CRs go through the webhook for
validation.

This also adds a new webhook rule to ensure the job template has a name set.

Signed-off-by: Matt Spilchen <matt.spilchen@vertica.com>