Fix timing that caused db add node before install
There is a small timing window with the operator where we would attempt
to do an add node before a pod has run install. This isn't a huge
problem as the add node would fail and do things in the correct order in
the subsequent reconcile iteration. But fixing this will cut down on the
amount of noise in the operator log.

The fix is to check for the install state when we run add node. If we
find a pod with a missing db and it doesn't have an install, we will
requeue rather than attempting the add node.