SSH secret mounting improvements
When using your own ssh keys through a ssh secret, we now mount the
secret for the root user too.

This comes in handy if you are running a hybrid k8s server and need to
run update_vertica from within the pod. Prior to this change you
couldn't do that because the ssh key for the root user was unrecognized.

This also mounts the ssh key without using sub path. This allows you to
update the ssh secret and see the change in the pod without having to
restart the pod.

Signed-off-by: Matt Spilchen <matt.spilchen@vertica.com>