New dispatcher abstraction (#416)
This adds a new abstraction for calls to admintools commands. The new
abstarction is in the vadmin package and the interface is called
Dispatcher. The purpose of this abstraction is to hide what we call to
handle the request. Currently, admin commands are all done through
admintools. But we are integrating with the vclusterops library, which
will be the alternative to admintools. Creating the correct dispatcher
struct will mean we either go through admintools or vclusterOps -- which
one depends on the feature flag (vertica.com/vcluster-ops="true"
annotation).

I only added create db and revive db in this PR. Subsequent ones will
add the rest of the commands.

The new abstraction uses the golang options pattern to implement all of
the various options passed into the API commands. I decided to create
separate packages for each option so that we don't have naming colisions
with the various With* functions added. So, you will see ones like
createdb and revivedb.

Both create and revive are mixed together in the init_db.go function. We
previously had separate functions for getting the command line arguments
to admintools and the actual admintools commands. Because we are moving
away from admintools, we now use the API options, which are then
converted to admintools command line options in the dispatcher.