Add dispatch abstraction for admintools -t re_ip (#422)
More work to hide admintools commands behind an abstraction so that we
can replace them with the vclusterOps library. This PR handles the
admintools -t re_ip command. There is setup for a map file that re_ip
needs. This was pushed into the abstraction too. When we create the
re_ip API for vclusterOps we won't be specifying the mapping in a file.