Implemented VClusterOps.CreateDB() (#430)
This is the integration work that calls vclusterops library to create a
database. When the vclusterops flag is true, we will create a
vclusterops dispatcher instead of an admintool dispatcher.

Since the test suite isn't ready, I didn't write an e2e test for
vclusterops create_db. The e2e test will be in another PR.