Fix flaky e2e test (#433)
I have seen the restart-with-liveness-probe test fail on occasion. One
part of the test will wait for the livenessProbe to kick in an restart
the pod. The test assertion check the lastState field in the Pod's
container status. I have seen this not be filled in as expected. I am
removing it from the assertion in an attempt to make the test run more
reliably.