Always run e2e-vcluster with latest master changes (#440)
The e2e-vcluster test suite needs to run with the latest version of the
server. For this reason, when we run that test suite we will use a
private image from the dockerhub repo with the lag lastest-master. We
also skip this test suite on forks because it requires secrets that are
only accessible through the vertica-kubernetes repo.

We still build the v2 image to make sure nothing is broken. But since it
uses an RPM from version 12.0.3 of the server, it doesn't have all of
the functionality that the e2e-vcluster test suite needs. So, no e2e
tests will run with it.