Use extra config parms for VClusterOps create DB (#441) (#442)
When calling create_db for admintools, we would create a file that has a
bunch of config options in it. When using the vclusterOps library, these
config parameters are passed as a map with the `ConfigurationParameters`
option. This change will set up that option. This will allow us to
create a database through vclusterOps with extra config parameters so
that we can connect to a communal backend like S3 or use advanced
options like S3 SSE.

Starts from #436