Fix e2e server upgrade tests
When run with the latest master, it runs with the 23.4.0 version. But
the CI will use the 12.0.2 image as the base image we are upgrading
from. This skips the 23.3.0 version, which violates the upgrade path.
We are now going to infer the base image to upgrade from using the
target image. This will allow us to run the e2e against 12.0.4 and
23.4.0 levels.