Fix the db existence check in the operator
This change fixes a bug in the operator's database existence check in
version 1.10.1. The operator used to check for the existence of the
catalog directory, but this could lead to errors if the create or revive
DB commands failed, as the catalog directory would remain even though
the database was not created or revived. This change fixes the bug by
looking for a specific file in the catalog directory that is only
present for fully initialized databases.