CSV updates to improve VerticaDB web ui creation (#466)
In OpenShift, you can create an instance of a VerticaDB using the web
UI. It builds the input based on annotations supplied for the various
fields in the CR. This updates those annotations to improve the UI for
an upcoming demo. Changes to v1beta1_verticadb.yaml were done because
the OpenShift web UI uses that to see default values.