Allow use of GSM for communal credential secret (#458)
This adds a feature flag to get the secret for the communal credential
from google secret manager (GSM) instead of from a k8s secret.

It is off by default. You must set the following annotation in the
VerticaDB CR:
```
kind: VerticaDB
metadata:
  name: v
  annotations:
    vertica.com/use-gcp-secret-manager: "true" 
spec:
  communal:
    credentialSecret: <full name of the secret in GSM>
...
```

The name of the GSM secret is taken from the existing
`spec.communal.credentialSecret` field.