Obfuscate access credentials in operator log
We have code that obfuscates access credentials before logging them in
the operator pod. This is looking for key/value pairs that we setup for
create or revive. The keys that we were searching for was recently
changed to be all lowercase. This caused the obfuscation code to not
find/replace the access key. Updating this to account for that key
change.

The prior release of the operator (1.11.1) did proper obfuscation. This
is a fix only for a regression in the current release.