Log successful add nodes in the event log (#467)
This fixes a small regression in this release. We were not logging an
event message when add nodes was successful.

Also, moving the pod facts invalidation for vclusterOps. Where it was
before meant that we would always invalidate the pod facts cache, even
for the case where the install reconciler does nothing.