Improve status retention in the status reconciler (#469)
This improves the status reconciler so that it doesn't lose state in the
VerticaDB's status fields. I saw a case where we were losing the vnode
information -- happened while working on the vclusterops integration and
was due to another integration bug. But I thought it would be good to
improve the resiliency of this reconciler to handle those edge cases.