Run reip/startdb with consistent node order (#471)
The node list that we pass to reip and start will now be in vnode sorted
order. Having a deterministic order for the pods will improve debugging.
Also, its temporarily needed for vclusterops integration as that library
has a strict requriement on the nodes being in vnode order. This will be
dropped later, but I think this improvement can stand on its own just
for the improvement to debugging.

This also includes some refactor to cleanup references to admintools in
the main vdb package since now the backend could be admintools or
vclusterops.