Fix e2e server upgrade test (#472)
We have to pick a slightly newer image for the
online-upgrade-block-downgrade test. Going back to an image that is too
old will fail to start due to permission problems required in the docker
entry point. This is only a problem with the nightly test that run with
the latest server code.