Allow use of GSM for superuser password secret (#474)
This is similar to #458 except we read from GSM for the superuser
password secret. We use the same feature flag as before. Here is a
sample CR to enable it.

```
kind: VerticaDB
metadata:
  name: v
  annotations:
    vertica.com/use-gcp-secret-manager: "true" 
spec:
  superuserPasswordSecret: <full name of the secret in GSM>
...
```

The name of the GSM secret is taken from the existing
`spec.superuserPasswordSecret` field.

Note, if GSM is being used we can't use the canary query for health probes
as that depends on the superuser password being a k8s secret. Instead,
we check if vertica is running by seeing if anyone is listening on the 
vertica client port.

---------

Co-authored-by: Matt Spilchen <matt.spilchen@microfocus.com>