Avoid using SQL to configure database after create (#473)
Traditionally when we create a database, we have run SQL commands to
initialize certain aspects of the database. We were able to do this
because the admintools -t create_db API has an option that let you
specify an SQL file to run. With the vclusterOps integration, it doesn't
have such a feature. Instead, the preferred way is to pass everything as
config parameters. This takes advantage of new config parameters being
added in 23.4.0.

The new config parameters are:
- InitialDefaultSubclusterName: Controls the name of the initial default
subcluster for new databases only. This has no effect if changed after
the database has been created. The operator previously controlled this
by doing an ALTER SUBCLUSTER SQL right after create
- InitialPreferredKSafe: Controls the initial value for preferred k-safe
for new databases. This has no effect if changed after the database has
been created. The operator previously controlled this by using the
'select set_preferred_ksafe()' meta-function after the database was
created.

Note: EncryptSpreadComm was enabled via SQL, but a config parm for it
always existed. However, it can only be set as a config parameter during
create db on versions 23.3.0 or newer. So, that will continue to be
enabled with SQL on older versions.

We cannot set EncryptSpreadComm=vertica for databases created with
vclusterOps. There is a bug that is preventing the nodes from coming up.
An internal ticket has been opened to fix the bug. In the mean time, I
have removed the use of that setting the vcluster e2e testcases.