Implement VClusterOps.RestartNodes (#483)
This integration incorporates the vclusterops restart nodes
functionality into the operator.

To get restart to work, I had to change the re-ip process. There were
cases where, when using admintools, we had to re-ip hosts that had been
installed but not yet added to the database. This was necessary to keep
the hosts' IP addresses up to date in admintools.conf. This step is not
necessary for vclusterops because there is no state file that needs to
be synchronized.