Include reconciler and vdb in log messages (#493)
Our logging library that we use (logr.Logger) has the ability to create
custom loggers with a name appended to it. This name is then displayed
for all logging calls done with it. This change makes use of that so we
can more easily trace where log messages are coming from.

We already include the VerticaDB name in each log messages. But we
weren't consistent with the logger we used, so some messages omitted it.

Here is a sample log message from before this change:
```
2023-09-07T16:30:19.188Z        INFO    controllers.VerticaDB   Reconcile client routing label     {"applyMethod": "Add"}
```

After this change, 3rd column now identifies the reconciler that logged
the message and we include the vdb.
```
2023-09-07T16:38:41.227Z        INFO controllers.VerticaDB.ClientRoutingLabelReconciler   Reconcile client routing label   {"verticadb": "spilly/vc", "applyMethod": "Add"}
```