Refresh with latest vcluster changes (#494)
New vcluster changes use a different structure for the logger called
vlog.Printer. It essentially wraps a logr.Logger with extra state to
know if we are printing to a file only. This is used to control the
logging for Print* messages; we avoid printing if we are already logging
to stdout.