Implement VClusterOps.DescribeDB (#495)
This adds to the vclusterops revive support by doing a prepass to
collect the cluster_config.json, then modifying local paths to match the
cluster config. We collect the cluster_config.json using the VReviveDB
call with the DisplayOnly option. This pre-existed for admintools based
deployments, but now we are hooking it up for vclusterops.

The format for vclusterops is slightly different. So, I updated the
reviveplanner package to have an admintools based parser and a
vclusterops based parser.

Both parser implement a new interface: ClusterConfigParser. The
admintools parser was moved to the pkg/reviveplanner/atparser package.
The vclusterops parser was added in the pkg/reviveplanner/vcparser
package.

The reviveplanner core logic works with either parser and was renamed
from ATPlanner to Planner to show that it isn't tied to the admintools
output.