Log event message for revive cluster lease error (#496)
This applies to the vclusterops implementation of revive. We will now
log an event message if we hit an cluster lease error. We use the new
ClusterLeaseNotExpired concrete type to know if we hit this error. This
gets us closer in sync with the admintools handling of specific errors
(see aterrors.go)