Fix e2e server upgrade tests in nightly runs (#500)
The server repo changed the version to 24.1.0. This caused issues in the
server upgrade e2e tests. In that test suite we guess the base image to
upgrade from. For nightly runs, we assume the base image is the
previously released image. The previously released image is 23.3.0. But,
in order to upgrade to 24.1.0 we need to start from 23.4.0. And since
23.4.0 isn't yet released, it broke the tests.

I am updating the guess script to use the private vertica-k8s repo so
that we do pick 23.4.0 for this case.

I had to change two testcases `offline-upgrade-ks-0` and
`offline-upgrade-ks-1` to use the base image for upgrades. It previously
had hard coded a vertica image (12.0.2). This works fine if upgrading to
23.x. But it violates upgrade path rules if you try to go directly to
24.x.

---------

Signed-off-by: Matt Spilchen <mspilchen@opentext.com>