Change operator to be cluster scoped
This will change the operator from being namespace scoped to cluster
scoped.

As part of this, a new reconciler had to be added to create the service
account to use with vertica pods. Previously, it would use the service
account that the operator uses. But this could be in a different
namespace now. This is a good change because it allowed us to scale back
the rbac rules the vertica pod has.

This removes some parameters from the helm chart that allowed you to
deploy the operator without cluster admin priv. The intention with a
cluster scoped operator that the helm chart will be run once by the
cluster admin. Then individual namespaces given to app developers who
want to use the operator.

Signed-off-by: Matt Spilchen <mspilchen@opentext.com>