Fix e2e-server-upgrade in nightly tests (#511)
It is currently failing because it cannot find the base vertica image
the tests start with. For nightly builds it is in a private repository.
We need to push that image into kind as part of test setup. Adding a new
make target for this. It will be called from within run-k8s-int-tests.sh
-- as part of `make docker-push`.