No longer ensures http service is running
The operator used to have a reconcile function to make sure that the
embedded http service is running. We are removing this because with the
new vclusterops deployments it must always be running. We plan on
changing the health probes to use an endpoint from the http service. So,
if the service becomes unhealthy the pod will be rescheduled. The
operator won't need to do anything on its end.

We considered leaving this reconciler around for admintools based
deployments, but there is very littly value in this outside of
vclusterops. So, it was simpliest just to remove that feature.