Deprecate httpServerMode CR parameter
We will use the deployment method when deciding how to configure vertica
for http access. All usages of httpServerMode have been removed. The
parameter has to stick around because that would break the API.

Renamed the e2e-http-server leg to e2e-leg-6. This only runs in
vclustermode now since that is the only deployment that will have the
http server setup.

Signed-off-by: Matt Spilchen <mspilchen@opentext.com>