Change scope of operator from namespace to cluster (#514)
This changes the scope of the operator from being namespace scoped to
cluster scoped.

This is quite a big change, especially in the e2e tests. Here are some
of the highlights:
- You can still can use make target to deploy/undeploy the operator as
usual. The change is the location of the operator. It is being deployed
in the verticadb-operator namespace by default, but you can override if
you set the NAMESPACE variable.
- We no longer support DEPLOY_WITH=random. You need to specify what type
of deployment you want (helm or olm). One of the test legs (e2e-leg-2)
was converted to run olm tests. The rest of the legs use helm.
- Since a single operator can control many VerticaDB instances, I use
additional concurrency when creating the operator. By default I use 5 go
routines meaning we can work on 5 different VerticaDB CRs at the same
time. This can be configured through new helm chart values.
- I got rid of some helm chart parameter that attempt to deal with allow
non-cluster admins to install the operator. This is no longer needed
because we can now deploy the operator once for the entire cluster. Gone
are the settings for serviceNameOverride and
skipRoleAndRoleBindingCreation. These are done unconditionally now when
deploying the operator.
- One challenge I had to deal with was how to get access to operator
config settings. Previously, the vertica pods had access to items from a
ConfigMap that told us the version of the operator and how the operator
was deployed (olm or helm). We still do this but because the ConfigMap
is in a different namespace, we had to surface this in a different
manner. The operator pod has this information as environment variables.
When the operator creates the pods, it will look at the environment
variables and create annotations based on their value. These annotations
are then stored as files in a known location using the downward API;
there is code in the server that fills in the dc_kubernetes_events table
with information from these known locations.
- Another challenge was the operator creates the serviceAccount (and
roles/rolebinding) for the vertica pod. Previously, we had just reused
the same serviceAccount that the operator uses. But since
serviceAccounts are namespace specific and the operator and the vertica
pods can be in different namespaces, we needed another way. As a future
item we will want the ability to specify a serviceAccount ahead of time.
For now the operator will always generate one if it is missing.
- Cleaned up some of the e2e legs. Gone is the e2e-enterprise leg. And I
added a new leg (e2e-leg-5) that will run all of the tests that need to
start the operator in a specific mode. The rest of the legs depend on
the operator to already be installed. Because of this, the new leg
cannot run with any test parallelism. OLM tests, which we use to run at
random will now only be run in e2e-leg-2. And I renamed the e2e-ezb test
suite to e2e-leg-1 in the CI

---------

Signed-off-by: Matt Spilchen <mspilchen@opentext.com>