Fix operator install issues during upgrade in vcluster mode (#517)
When trying to upgrade and scale up at the same time, the operator got
stuck at `OfflineUpgradeReconciler.InstallReconciler` preventing the
operator from moving to the next actors to perform the scale up. To fix
that, we relaxed `InstallReconciler` to not wait for all pods (in
podfacts) to do the install before moving to the next actor.