Move directory prep into dispatcher abstraction (#510)
This PR relocates the 'directory prep' functionality from
`pkg/controllers/vdb/paths.go` to the dispatcher abstraction. The
'directory prep' is only used by AT and not by vclusterops. The key
benefit here is that it moves some ugly sudo calls into code that is
specific to admintools deployments. This gets us a step closer to
removing the requirement of sudo access for vclusterops deployments.