Add v1 API for VerticaDB (#519)
This is the first phase in introducing a v1 API for VerticaDB. The old
version, v1beta1, is deprecated, but will continue to work for now.
However, you will receive a deprecation warning each time you use it. In
a future release, v1beta1 will be removed altogether.

```
$ kubectl apply -f vdb.yaml
Warning: vertica.com/v1beta1 VerticaDB is deprecated, use vertica.com/v1 VerticaDB
verticadb.vertica.com/vc configured
```

In this PR, the two versions of the API are identical except for one
field: spec.communal.hadoopConfig has been moved to spec.hadoopConfig.
This change was made to prove that the conversion webhook is working
properly. Follow-on work will move and rename other fields in the v1
API.

This PR also adds infrastructure to support two versions. A new
conversion webhook has been added that will automatically convert
between the v1beta1 and v1 versions. Subsequent PRs will start to make
changes to the v1 API.

Note that you can still deploy the operator without a webhook. However,
if you do, you must use the v1 API immediately. The VerticaDB controller
now works exclusively with the v1 API, so you must use that version.

About half of the e2e testcases have been modified to use the new V1
API.

The CA bundle for the webhook now needs to be included in the CRD, since
that is where the conversion webhook is configured. The operator will
update this at runtime to handle the various ways the webhook TLS can be
specified. The helm chart parameter webhook.caBundle has been removed,
since it was already deprecated and there is little value in using the
bundle specified there and applying it to the CRD conversion webhook
config.