Pass list of expected node names to vcluster VAddNode API (#520)
A new vclusterops change allows the caller to pass in a list of expected
node names for VAddNodes. This allows vclusterops a chance to cleanup a
partially added node if it wasn't included in the list. This PR will
move up the VCluster API and fill in this new parameter to better handle
add node failures.