Allow the ServiceAccount to be specified for vertica server pods (#521)
This allows the serviceAccount to be specified for the vertica server
pods. Now that the operator runs at cluster scope, we have separated the
service account used to run the operator with the one to run the vertica
server pods. In #514 the operator will generate the service account,
along with a role and rolebinding, to use. This takes it one step
further and allow the user to create the service account ahead of time.

This allows for better integration where the service account needs to be
created ahead of time to setup access (e.g. IAM profile via AWS IRSA).

The vertica server pod needs to have read access to k8s secrets in order
to read certs. So, a role and rolebinding must be attached to the
service account.

One way that someone could use this:
1. Create the ServiceAccount. If setting up IRSA, this step could
include adding an AWS annotation so that the serviceAccount has access
to a particular IAM profile.
`kubectl create serviceaccount vertica-server-sa`
2. Create the role that the vertica server needs:
`kubectl create -f config/rbac/vertica-server-role.yaml # This will be
available as a release artifact too`
3. Create the rolebinding to link the two:
`kubectl create rolebinding vertica-server-rolebinding --role
vertica-server-role --serviceaccount=<namespace>:<serviceAccountName>`
4. Create the VerticaDB and fill in the `spec.serviceAccountName` field
with the name of the service account (e.g. vertica-server-sa)

Note, if the service account name is specified, then we assume the
necessary role and rolebindings are setup too. The operator will only
generate those if it generates the service account also.