Add soak testing framework (#525)
This brings back the soak test we had before. The soak test will test
the operator over a long interval. It splits the test into multiple
iterations. Each iteration generates a random workload that is comprised
of pod kills and scaling. At the end of each iteration, the test waits
for everything to come up. If the test is successful, it proceeds to
another iteration. It repeats this process for a set number of
iterations or indefinitely.

The tests in an iteration are run through kuttl. The random test
generation is done by the kuttl-step-gen tool.

Here are the steps needed to run this test.

1. Create the databases that you want to test.
2. Create a config file to outline the databases to test and how you
want the test framework to react. A sample one can be found in
tests/soak/soak-sample.cfg.
```
$ cp tests/soak/soak-sample.cfg local-soak.cfg
$ vim local-soak.cfg
```
3. Decide on the number of iterations you would like to run:
```
$ export NUM_SOAK_ITERATIONS=10  # Can use -1 for infinite
```
4. Kick off the run.
```
$ make run-soak-tests
```

---------

Signed-off-by: Matt Spilchen <mspilchen@opentext.com>