VCluster API Refresh (#526)
This picks up the latest vclusterops package. The key change is it will
hide sensitive config parms from the various logs. Originally, it was to
use a new parameter to pass down config parms to create DB, but after a
few revisions we opted to reuse the existing field, so this just becomes
a refresh of the package and no API changes.

Co-authored-by: ctran <ctran@microfocus.com>