Fix intermittent failure in revivedb-multi-sc (#529)
We are seeing intermittent failures in the revivedb-multi-sc testcase
when run in vclusterops. The test creates a 2 subcluster database. The
test waits in step 20 for the primary and then waits for the other
subcluster to come up in step 25. I suspect what is happening is that we
are bringing the secondary subcluster up so fast that the check at step
20 doesn't get a chance the intermediate state. So, I am combining step
20 and 25 to close this gap.