Rename v1 API fields (#531)
Since we have not yet released the v1 API, we are free to change it. I
am making the following changes to the API:
- removed spec.ignoreClusterLease in favour of an annotation:
vertica.com/ignore-cluster-lease
- removed spec.ignoreUpgradePath in favour of an annotation:
vertica.com/ignore-upgrade-path
- removed spec.RestartTimeout in favour of an annotation:
vertica.com/restart-timeout
- changes spec.temporarySubclusterRouting to be a pointer. This allows
us to check for nil to see if this setting is used. And it won't show a
default when you list the YAML manifest of a CR.

The v1beta1 API wasn't changed. The conversion webhook has been updated
to convert between the two versions.