Skip restart reconciler when DB init fails (#527)
This is a bug fix for vclusterops deployments. When create_db or
revive_db fails, the re-ip process in the restart_reconciler will be
trigged and also fail. The solution is to skip the restart reconciler so
that the create/revive can be attempted again. We will skip based on the
DBInitialized status condition.

---------

Co-authored-by: cchen-vertica <cai.chen@microfocus.com>
Co-authored-by: ctran <ctran@microfocus.com>