Ensure http service starts for vcluster create db
This fixes a timing issue that could cause the vclusters create db
operation to fail. This can happen if the install is not run before the
create db. The installer sets up a configuration file that starts the
HTTP service. If this is skipped, the HTTP service will not start and
create db will not succeed because it needs that endpoint to communicate
with.