Allow spread encryption for vclusterops deployments (#536)
If spread encryption is selected in the CR, we now allow it for
vclusterops deployments. It takes advantage of a new NMA endpoint to
copy the key into the spread.conf.

I included with this change the removal of a cluster restart when
enabling spread encryption. This was first available in 23.3.0. This
change applies to admintools deployments as well as vclusterops. Older
versions of the server will continue to restart the cluster.

Closes #535