Rename v1 API fields: kSafety (#537)
Same rationale as in #531. This moves the `spec.kSafety` field to an
annotation. The v1beta1 API is not impacted by this.

While developing this change I came across an issue that showed the
conversion webhook wasn't being used in olm deployments. The CRD, where
the conversion webhook is configured, was left untouched. So, I am
including a fix for that. We now copy the webhook CA cert into the CRD
for olm based deployments.