Set a custom logger for each vcluster API
This creates a custom logger for each vcluster API call. The purpose is
so that we can construct a logr.Logger object using the WithName()
function. This will give us a better idea of where the various log
entries are coming from. For instance, with this change, you can see the
3 log messages below all originate from the CreateDB API call (3rd
column in each log statement).

2023-10-14T21:14:35.123Z    INFO    controllers.VerticaDB.CreateDB Starting vcluster CreateDB  {"verticadb": "kuttl-test-able-insect/v-client-access"}
2023-10-14T21:14:35.123Z    INFO    controllers.VerticaDB.CreateDB Setup create db options {"verticadb": "kuttl-test-able-insect/v-client-access", "hosts": "10.244.0.28"}
2023-10-14T21:14:35.123Z    INFO    controllers.VerticaDB.CreateDB starting VCreateDatabase    {"verticadb": "kuttl-test-able-insect/v-client-access"}

Signed-off-by: Matt Spilchen <mspilchen@opentext.com>