Rename v1 API fields: kerberosRealm and kerberosServiceName (#544)
Another change to move CR parms to annotations in the v1 API. This
removes both kerberosServiceName and kerberosRealm. Both of them can be
specified through the `spec.communal.additionalConfig`, so it didn't
make sense to have separate fields. See #531 to understand why we are
changing the v1 API. No change is done to the v1beta1 API.