Rename v1 API fields: add subcluster type (#547)
This adds a string field called Type to subclusters. This is more
expressive as it replaces the isPrimary and isTransient boolean. It also
allows for future expansion if we introduce different subcluster types
in the future. See #531 to understand why we are making this change. No
change is done to the v1beta1 API.