Rename v1 API fields: status field updates (#548)
This updates the status fields for the v1 API. A summary of the changes
are:
- Rename ImageChangeInProgress condition variable to UpdateInProgress
- Remove the ReadOnly state, since it no longer needs to be surfaced
- Remove the install count at the cluster and subcluster level. We still
track install per pod, but use functions to sum up the count at higher
levels

I am also updating the print columns of the VerticaDB when you do get
vdb. Sample output looks like this:
```
$ kubectl get vdb
NAME   AGE   VERSION            READY   SUBCLUSTERS
vc     56s   v24.1.0-20231012   3/3     1
```

See #531 to understand why we are making this change. No change is done
to the v1beta1 API.