Include scrutinize storage in PV
For vclusterops deployments, we want to carve out a portion of the PV
for use with scrutinize. The directory will be /tmp/scrutinize. We are
doing this because we don't want to risk filling up the node's local
directory with compressed diagnostics. These can be quite big. If the
node's space is used then k8s will simply reschedule the pod, forcing
the scrutinize to start over.