VCluster API Refresh: logging update (#553)
This picks up the latest vclusterops package. The key change is the
logging for the library now uses the logr.Logger API everywhere. So, all
of the logging for the operator and vclusterops should now follow the
same format.

I also made a CI change. I noticed that we were having trouble
downloading kubectl. This should be already included in the GitHub
runners so removing the download step.